SimPad Touch New Version

+++ src/debounce.h
@@ -4,7 +4,7 @@
 #include "ch552.h"
 #include "bsp.h"
 
-typedef uint8_t debounce_type;
+typedef uint16_t debounce_type;
 #define DEBOUNCE_SIZE KEY_COUNT
 
 void debounceInit();
